refactor(prototyper): split platform boot entry points into boot.rs
Move the five safe boot entry points over PLATFORM from platform/mod.rs
into a re-exported platform/boot.rs submodule; call sites are unchanged.
